Backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sentence-transformers
faiss-cpu
python-dotenv
google-genai
pypdf
numpy